Pushing using long ints instead of ints

+++ 0x13-bit_manipulation/1-print_binary.c
@@ -10,7 +10,7 @@
 
 void print_binary(unsigned long int n)
 {
-	unsigned int i, temp, cntr;
+	unsigned long int i, temp, cntr;
 
 	cntr = 0;
 	i = 1;
